backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy
passlib==1.7.4
python-jose==3.3.0
python-multipart
python-dotenv==1.0.0
openai==0.28.0
tenacity==8.2.3
uvicorn==0.24.0
asgiref==3.7.2
